+++ packages/flutter_adaptive_cards_fs/lib/src/cards/elements/rating.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_adaptive_cards_fs/src/adaptive_mixins.dart';
 import 'package:flutter_adaptive_cards_fs/src/additional.dart';
-import 'package:flutter_adaptive_cards_fs/src/riverpod_providers.dart';
 import 'package:flutter_adaptive_cards_fs/src/utils/utils.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class AdaptiveRating extends StatefulWidget with AdaptiveElementWidgetMixin {
   AdaptiveRating({
@@ -40,9 +38,7 @@ class AdaptiveRatingState extends State<AdaptiveRating>
 
   @override
   Widget build(BuildContext context) {
-    final resolver = ProviderScope.containerOf(
-      context,
-    ).read(styleReferenceResolverProvider);
+    final resolver = styleResolver;
     Color starColor;
     if (color == 'marigold') {
       starColor =