+++ packages/flutter_adaptive_cards_fs/lib/src/cards/inputs/text.dart
@@ -2,9 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_adaptive_cards_fs/src/adaptive_mixins.dart';
 import 'package:flutter_adaptive_cards_fs/src/additional.dart';
-import 'package:flutter_adaptive_cards_fs/src/riverpod_providers.dart';
 import 'package:flutter_adaptive_cards_fs/src/utils/utils.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 ///
 /// httfps://adaptivecards.io/explorer/Input.Text.html
@@ -94,9 +92,7 @@ class AdaptiveTextInputState extends State<AdaptiveTextInput>
                     borderSide: BorderSide(width: 1),
                   ),
                   filled: true,
-                  fillColor: ProviderScope.containerOf(context)
-                      .read(styleReferenceResolverProvider)
-                      .resolveInputBackgroundColor(
+                  fillColor: styleResolver.resolveInputBackgroundColor(
                         context: context,
                         style: null,
                       ),