+++ packages/flutter_adaptive_cards_fs/CHANGELOG.md
@@ -5,9 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.0] - 2026-04-19
+## [0.5.0]
 
 - version numbers were sync'd to 0.5.0
+- Added `ChartColorsConfig` to `HostConfig` to allow custom color palettes for charts.
+- Added `ReferenceResolver` with `resolveChartPalette` and `resolveChartColor` methods.
 - migrated more hard coded styles to the resolver
 
 ## [0.4.0] - 2026-04-14